(editor) Extract PollenEditorBase with shared section methods
Move the reusable editor pieces (translation, integration dropdown,
allergen-name resolver, silam-weather check, deepMerge, and the
Integration/Location, Allergens and Appearance section renderers) into
a PollenEditorBase class. The card editor now extends it and calls the
shared section methods, so the badge editor can reuse the same sections
without duplicating any code. No behavior change to the card editor.
